Move vault-related methods into environment.secrets_backend
