+++ src/commcare_cloud/commands/ansible/ansible_playbook.py
@@ -128,7 +128,7 @@ def run_ansible_playbook(
 
         ask_vault_pass = public_vars.get('commcare_cloud_use_vault', True)
         if ask_vault_pass:
-            cmd_parts += ('--vault-password-file={}/echo_vault_password.sh'.format(ANSIBLE_DIR),)
+            cmd_parts += environment.secrets_backend.get_extra_ansible_args()
 
         cmd_parts_with_common_ssh_args = get_common_ssh_args(environment, use_factory_auth=use_factory_auth)
         cmd_parts += cmd_parts_with_common_ssh_args
@@ -140,7 +140,7 @@ def run_ansible_playbook(
         return subprocess.call(cmd_parts, env=env_vars)
 
     def run_check():
-        with environment.suppress_vault_loaded_event():
+        with environment.secrets_backend.suppress_vault_loaded_event():
             return ansible_playbook(environment, playbook, '--check', *unknown_args)
 
     def run_apply():
