+++ src/commcare_cloud/environment/main.py
@@ -1,20 +1,15 @@
-import getpass
 import os
 import re
-import sys
+
 from collections import Counter
-from contextlib import contextmanager
 from datetime import datetime
 
-import datadog.api
+
 import yaml
 from ansible.inventory.manager import InventoryManager
 from ansible.parsing.dataloader import DataLoader
-from ansible.parsing.vault import AnsibleVaultError
 from ansible.vars.manager import VariableManager
-from ansible_vault import Vault
 from memoized import memoized, memoized_property
-from six.moves import shlex_quote
 
 from commcare_cloud.environment.constants import constants
 from commcare_cloud.environment.exceptions import EnvironmentException
@@ -28,13 +23,13 @@ from commcare_cloud.environment.schemas.postgresql import PostgresqlConfig
 from commcare_cloud.environment.schemas.proxy import ProxyConfig
 from commcare_cloud.environment.schemas.terraform import TerraformConfig
 from commcare_cloud.environment.schemas.prometheus import PrometheusConfig
+from commcare_cloud.environment.secrets.backends.ansible_vault.main import AnsibleVaultSecretsBackend
 from commcare_cloud.environment.users import UsersConfig
 
 
 class Environment(object):
     def __init__(self, paths):
         self.paths = paths
-        self.should_send_vault_loaded_event = True
 
     @property
     def name(self):
@@ -88,93 +83,23 @@ class Environment(object):
             )
 
     def get_ansible_vault_password(self):
-        """Get ansible vault password
-
-        This method has a side-effect: it records a Datadog event with
-        the commcare-cloud command that is currently being run.
-        """
-        self.get_vault_variables()
-        return self._get_ansible_vault_password()
+        return self.secrets_backend.get_ansible_vault_password()
 
     def get_vault_variables(self):
-        """Get ansible vault variables
-
-        This method has a side-effect: it records a Datadog event with
-        the commcare-cloud command that is currently being run.
-        """
-        vault_vars = self._get_vault_variables()
-        if "secrets" in vault_vars:
-            self.record_vault_loaded_event(vault_vars["secrets"])
-        return vault_vars
-
-    @memoized
-    def _get_ansible_vault_password(self):
-        return (
-            os.environ.get('ANSIBLE_VAULT_PASSWORD') or
-            getpass.getpass("Vault Password for '{}': ".format(self.name))
-        )
-
-    @memoized
-    def _get_vault_variables(self):
-        # try unencrypted first for tests
-        with open(self.paths.vault_yml, 'r') as f:
-            vault_vars = yaml.safe_load(f)
-        if isinstance(vault_vars, dict):
-            return vault_vars
-
-        while True:
-            try:
-                vault = Vault(self._get_ansible_vault_password())
-                with open(self.paths.vault_yml, 'r') as vf:
-                    return vault.load(vf.read())
-            except AnsibleVaultError:
-                if os.environ.get('ANSIBLE_VAULT_PASSWORD'):
-                    raise
-                print('incorrect password')
-                self._get_ansible_vault_password.reset_cache(self)
+        return self.secrets_backend.get_vault_variables()
 
     def get_vault_var(self, var):
-        path = var.split('.')
-        context = self.get_vault_variables()
-        for node in path:
-            context = context[node]
-        return context
+        return self.secrets_backend.get_vault_var(var)
+
+    def record_vault_loaded_event(self, secrets):
+        return self.secrets_backend.record_vault_loaded_event(secrets)
+
+    @memoized_property
+    def secrets_backend(self):
+        return AnsibleVaultSecretsBackend(self.name, self.paths.vault_yml)
 
     def get_ansible_user_password(self):
         return self.get_vault_variables()['ansible_sudo_pass']
-
-    def record_vault_loaded_event(self, secrets):
-        if (
-            self.should_send_vault_loaded_event and
-            secrets.get('DATADOG_API_KEY') and
-            self.public_vars.get('DATADOG_ENABLED')
-        ):
-            self.should_send_vault_loaded_event = False
-            datadog.initialize(
-                api_key=secrets['DATADOG_API_KEY'],
-                app_key=secrets['DATADOG_APP_KEY'],
-            )
-            datadog.api.Event.create(
-                title="commcare-cloud vault loaded",
-                text=' '.join([shlex_quote(arg) for arg in sys.argv]),
-                tags=["environment:{}".format(self.name)],
-                source_type_name='ansible',
-            )
-
-    @contextmanager
-    def suppress_vault_loaded_event(self):
-        """Prevent "run event" from being sent to datadog
-
-        This is only effective if `self.get_vault_variables()` has not
-        yet been called outside of this context manager. If it has been
-        called then the event has already been sent and this is a no-op.
-        """
-        value = self.should_send_vault_loaded_event
-        self.should_send_vault_loaded_event = False
-        try:
-            yield
-        finally:
-            self.should_send_vault_loaded_event = value
 
     @memoized_property
     def public_vars(self):
