+++ src/commcare_cloud/manage_commcare_cloud/list_vault_keys.py
@@ -29,7 +29,7 @@ class ListVaultKeys(CommandBase):
         for env in envs:
             print('[{} (blank to skip)] '.format(env), end='')
             environment = get_environment(env)
-            passwd = environment._get_ansible_vault_password()
+            passwd = environment.secrets_backend._get_ansible_vault_password()
             if passwd:
                 var_keys[env] = dict(get_flat_list_of_keys(environment.get_vault_variables()))
